Experimental fix for non-threaded rendering performance/ frame sync handling, Part 2:
Reimplement frame finish for app-node with non-threaded pipes. In this
case, the app-node finishes the previous frame before the first blocking
operation of the current frame. If there are no non-threaded app-node
pipes, or no dependent operations, the execution is not changed.

Remove priority packets and early node finish.

Disable OpenMP on gcc-4.2

[x] May break build
[ ] Breaks existing applications
[ ] Bugfix
[x] New Feature
[ ] Cleanup
[ ] Optimization
[ ] Documentation