Address the possibility of lost data on disconnect.
Address a possible chance that while the event thread has exited
because the connection was dropped, there could still be data
available in the sink buffer for the application to read.  Let
the application handle closing the eventfd after all available
bytes have been read.

Also change the behavior of failing to send an ACK/flow control
message to WARN instead of close/fail as we could be in the
above mentioned state and want that data.

[ ] May break build
[ ] May break existing applications (see CHANGES.txt)
[x] Bugfix
[ ] New Feature
[ ] Cleanup
[ ] Optimization
[ ] Documentation