Closes 109: Race and performance issue with multicast object distribution
Bug: in certain conditions, instance packets where unicasted and reply
     packets multicasted, which is a race.
Opt: reply packets only need to be multicasted if instance data was multicasted

Implement by pushing the send of the success/reply packets down to the
change managers, which know if instance data needs to be send. Refactor
addSlave, which does all this magic, to remove C&P code in existing CM's.

[ ] May break build
[ ] May break existing applications (see CHANGES.txt)
[x] Bugfix
[ ] New Feature
[x] Cleanup
[x] Optimization
[ ] Documentation