Fixed two issues with RDMA connection establishment.
In some systems rdma_getaddrinfo returns a list of addresses where
the first is from the AF_IB familiy. Apparently this only happens
in the client and the server, which binds using AF_INET[6] rejects
the connection.

In route resolution it is necessary to wait for the event
RDMA_CM_EVENT_ROUTE_RESOLVED fter using rdma_set_option when the
rdma_addrinfo struct contains routing information.