Workaround a compilation error related to math headers.
This is a workaround for the cmath/math.h mess with gcc vs clang.
Lunchbox replaced cmath with math.h in a header to fix compilation
errors in OS/X 10.13 when cmath is included. This in turn has caused
test/issue38.cpp to stop compiling because the proper std::abs are
missing, however including cmath in 10.13 is causing compilation errors
for unknown reasons.